+++ docs/news.rst
@@ -1,6 +1,20 @@
 News
 ====
 
+v5.2 (2020-01-07)
+-----------------
+
+Lab
+^^^
+* Use line buffering for run.err files.
+
+Downward Lab
+^^^^^^^^^^^^
+* Preserve line breaks for error logs in tables.
+* If an error log in a table has more than 100 lines, omit surplus lines from the middle of the log.
+* Always print the number of runs with unexplained errors when generating any type of report.
+
+
 v5.1 (2019-12-10)
 -----------------